refactor(yield): extract YieldPageRedirect into reusable TradeRouteRedirect
