+++ apps/cowswap-frontend/src/modules/trade/index.ts
@@ -20,6 +20,7 @@ export * from './utils/tradeFlowAnalytics'
 export * from './utils/logger'
 export { addPendingOrderStep } from './utils/addPendingOrderStep'
 export * from './const/tradeUrl'
+export * from './containers/TradeRouteRedirect'
 export * from './containers/TradeWidget'
 export * from './containers/TradeConfirmModal'
 export * from './containers/TradeWidgetLinks'
